creating authentication and logins folder
