+++ src/Router/router.ts
@@ -1,6 +1,8 @@
 import { Router } from "express";
 import { UserController } from "../Controllers/User/userController";
 import { AdminController } from "../Controllers/Admin/admincontroller";
+import { LoginController } from "../Auth/LoginUsers/authLoginUser";
+import { LoginAdmincontroller } from "../Auth/LoginAdmin/authLoginAdmin";
 
 const router = Router();
 
@@ -8,13 +10,18 @@ const userController = new UserController();
 
 const adminController = new AdminController();
 
+const loginUsercontroller = new LoginController();
+
+const loginAdminController = new LoginAdmincontroller();
+
+
 //Rotas de Usuarios
 router.get("/users/:id", userController.listUsers);
 router.post("/users/create", userController.createUser);
 router.delete('/users/delete/:id', userController.deleteUser);
 
 //Login Usuarios
-router.post('/users/login',userController.login);
+router.post('/users/login', loginUsercontroller.login);
 
 
 //Rotas de Admin
@@ -23,6 +30,6 @@ router.post('/admin/create',adminController.createAdmin);
 router.delete('/admin/delete/:id', adminController.deleteAdmin);
 
 //Login Admin
-router.post('/admin/login',adminController.loginAdmin);
+router.post('/admin/login',loginAdminController.loginAdmin);
 
 export default router;
